+++ src/encoder_options.rs
@@ -1,4 +1,4 @@
-use std::fmt::Debug;
+use std::{fmt::Debug, num::NonZeroU64};
 
 #[cfg(feature = "ppmd")]
 use ppmd_rust::{PPMD7_MAX_MEM_SIZE, PPMD7_MAX_ORDER, PPMD7_MIN_MEM_SIZE, PPMD7_MIN_ORDER};
@@ -36,17 +36,15 @@ impl LZMAOptions {
 #[derive(Debug, Clone)]
 /// Options for LZMA2 compression.
 pub struct LZMA2Options {
-    pub(crate) options: lzma_rust2::LZMAOptions,
+    pub(crate) options: lzma_rust2::LZMA2Options,
     pub(crate) threads: u32,
-    pub(crate) stream_size: Option<u64>,
 }
 
 impl Default for LZMA2Options {
     fn default() -> Self {
         Self {
-            options: lzma_rust2::LZMAOptions::with_preset(6),
+            options: lzma_rust2::LZMA2Options::with_preset(6),
             threads: 1,
-            stream_size: None,
         }
     }
 }
@@ -60,9 +58,8 @@ impl LZMA2Options {
     /// * `level` - Compression level (0-9, clamped to this range)
     pub fn from_level(level: u32) -> Self {
         Self {
-            options: lzma_rust2::LZMAOptions::with_preset(level),
+            options: lzma_rust2::LZMA2Options::with_preset(level),
             threads: 1,
-            stream_size: None,
         }
     }
 
@@ -72,23 +69,23 @@ impl LZMA2Options {
     /// # Arguments
     /// * `level` - Compression level (0-9, clamped to this range)
     /// * `threads` - Count of threads used to compress the data
-    /// * `stream_size` - Size of each independent stream of uncompressed data.
+    /// * `chunk_size` - Size of each independent chunk of uncompressed data.
     ///   The more streams can be created, the more effective is
     ///   the multi threading, but the worse the compression ratio
-    ///   will be (will be clamped to be at least 256 KiB).
-    pub fn from_level_mt(level: u32, threads: u32, stream_size: u64) -> Self {
-        Self {
-            options: lzma_rust2::LZMAOptions::with_preset(level),
-            threads,
-            stream_size: Some(stream_size.max(lzma_rust2::MIN_STREAM_SIZE)),
-        }
+    ///   will be (value will be clamped to have at least the size of the dictionary).
+    pub fn from_level_mt(level: u32, threads: u32, chunk_size: u64) -> Self {
+        let mut options = lzma_rust2::LZMA2Options::with_preset(level);
+        options.set_chunk_size(NonZeroU64::new(
+            chunk_size.max(options.lzma_options.dict_size as u64),
+        ));
+        Self { options, threads }
     }
 
     /// Sets the dictionary size used when encoding.
     ///
     /// Will be clamped between 4096..=4294967280.
     pub fn set_dictionary_size(&mut self, dict_size: u32) {
-        self.options.dict_size =
+        self.options.lzma_options.dict_size =
             dict_size.clamp(lzma_rust2::DICT_SIZE_MIN, lzma_rust2::DICT_SIZE_MAX);
     }
 }
@@ -582,7 +579,7 @@ impl EncoderOptions {
             #[cfg(feature = "compress")]
             EncoderOptions::LZMA(o) => o.0.dict_size,
             #[cfg(feature = "compress")]
-            EncoderOptions::LZMA2(o) => o.options.dict_size,
+            EncoderOptions::LZMA2(o) => o.options.lzma_options.dict_size,
             #[allow(unused)]
             _ => 0,
         }